requirements.txt
django==3.2
uwsgi==2.0.20
python-dotenv==0.21.0
django-split-settings==1.2.0
django-debug-toolbar==3.4.0
django-cors-headers==3.13.0
psycopg2==2.9.5
pre-commit==2.15.0
black==22.10.0
flake8==5.0.4
flake8-black
flake8-per-file-ignores